+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## UNRELEASED
+
+- Improved serialization for Django forms and models: pickle now correctly handles unsaved Django model instances and validates bound forms during unpickling.
+
 ## 1.12.3 (2024-08-14)
 
 - Created populate_form_with_data() utility function to populate a form with data from the user input.